finished all the big things
completed the gallery so that the images all fit and look good using grids., then went on to complete the colors that may or not be kept or deleted, and finally came about creating the contact form that is a clickable popup. All that is left for now is the little changes and bug fixes along with making sure mobile is working correct.